requirements.txt
numpy
torch
torchvision
pandas
Pillow
onnx
onnxruntime
onnxscript
